chore: tidy up after the storage rework
Four cleanups found while working through the storage changes, none of which
alter behaviour:

- Document that extractSessionData() hands back this.data.sessions by
  reference. Both pickKeys() and normalizeSessionData() copy shallowly, so it
  is a view rather than a snapshot. Stripping history from it in place would
  have silently wiped the history the UI was showing; the note is at the
  definition so the next caller does not have to rediscover that.

- Rename applyDefaultSettingsToCurrentScope() to applyDefaultSettings(). The
  "scope" was the vault-local settings layer, which no longer exists.

- Drop /main.js from .gitignore. The file is tracked, so the line never did
  anything except suggest it was not.

- Collect the six copies of the obsidian module stub into tests/helpers.js.
  It also resolves an i18n locale, which every copy had to remember to do
  because persistence.js reads i18n.L.

The unused-key warnings from check:i18n were left alone: it only matches
static L.key access, and roughly twenty of the keys it reports are reached
through L[labelKey] in the status bar settings. Deleting on its say-so would
have broken that UI.
